requirements.txt
# Data Lake Explorer - Snowflake Data Access Visualization
# Python 3.11+ recommended

# Core Streamlit framework
streamlit>=1.28.0

# Data processing
pandas>=2.0.0

# Visualization
plotly>=5.0.0
pyvis>=0.3.2
networkx>=3.0

# Image processing
pillow>=10.0.0

# Snowflake connectivity (optional - for live data)
# Comment out if running locally with sample data only
snowflake-snowpark-python>=1.11.0

